pyproject.toml
[project]
name = "afo"
version = "0.1.0"
description = "Autonomous Firewall Orchestrator - Natural language to nftables rules"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=0.4.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
